// interfaces/User/User.ts
export type IProfileResponse = {
    count_comment: number;
    count_sukien: number;
    count_view: number;
    device_register: string;
    email: string;
    id_user: number;
    ip_register: string;
    link_avatar: string;
    user_name: string;
}
export type IResponseLogin = {
    count_comment: number;
    count_sukien: number;
    count_view: number;
    device_register: string;
    email: string;
    id_user: number;
    ip_register: string;
    link_avatar: string;
    message: string;
    status: number;
    time_coin: number;
    token: string;
    user_name: string;
}